pyproject.toml
[project]
name = "task-management-api"
version = "0.1.0"
description = "a simple task management api created using fastapi"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic==1.4.1",
    "fastapi[standard]==0.115.8",
    "psycopg2-binary>=2.9.10",
    "python-dotenv==1.0.0",
    "ruff==0.9.7",
    "sqlalchemy==2.0.38",
    "uvicorn==0.34.0",
]
